refactor: ensure mixins do not create shared properties in AbstractGraph (#879)
The current implementation of the function applying mixins creates
shared state for "object" or "array" properties.

For such properties, ensure they are all defined in AbstractGraph and
add tests to validate that there is no global
states. Here, this includes a fix for 'alternateEdgeStyle' which is no
longer a global state.

In some mixin type files, remove properties that are already defined in
AbstractGraph.