refactor!: move tooltip methods to TooltipHandler (#640)
Move `getTooltip` and `getTooltipForCell` from `AbstractGraph` to the `TooltipHandler` plugin as this is the only place where they are used.
This improves modularity and tree-shaking, clarifies responsibilities, and reduces the amount of code in `AbstractGraph` by removing the `TooltipMixin`.

The stories have been updated accordingly and can be used as examples for adapting application code.

BREAKING CHANGES: The `getTooltip` and `getTooltipForCell` methods have been moved from `AbstractGraph` to the `TooltipHandler` plugin.
If you were overriding these methods in a `AbstractGraph` subclass, you should now extend `TooltipHandler` instead.